requirements.txt
numpy>=1.19
toml>=0.10
numba>=0.51
scipy>=1.5
svg.path>=6.2